requirements.txt
pytorch
neptune-client
torchdata
torchtext
